+++ db/migrate/20170408165745_create_certifications.rb
@@ -3,7 +3,7 @@ class CreateCertifications < ActiveRecord::Migration[5.0]
     create_table :certifications do |t|
       t.date :date
       t.references :contract, foreign_key: true
-      t.decimal :total
+      t.integer :total
       t.timestamps
     end
   end
