+++ db/migrate/20170408173356_create_certification_details.rb
@@ -3,11 +3,11 @@ class CreateCertificationDetails < ActiveRecord::Migration[5.0]
     create_table :certification_details do |t|
       t.references :certification, foreign_key: true
       t.references :employee, foreign_key: true
-      t.integer :quantity, default: 0
-      t.decimal :pay, default: 75558
-      t.decimal :discount, default: 0
+      t.float :quantity, default: 0
+      t.integer :pay, default: 75558
+      t.integer :discount, default: 0
       t.string :observation
-      t.decimal :subtotal, default: 0
+      t.integer :subtotal, default: 0
 
       t.timestamps
     end
